move get_clusters to utils/cluster.py
